+++ zeebe/protocol-impl/src/main/java/io/camunda/zeebe/protocol/impl/record/value/authorization/AuthorizationOwnerType.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.protocol.impl.record.value.identity;
+package io.camunda.zeebe.protocol.impl.record.value.authorization;
 
 public enum AuthorizationOwnerType {
   USER,
